Restructure with separate files for Data and stats pages, folder for styles, and types file for shared types
