+++ lib/components/Inputs.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import Form from "react-bootstrap/Form";
 import Creatable from "react-select/creatable";
-import getStyles from "./styles";
-
-type OptionType = { label: string; value: string };
+import selectStyles from "../utils/selectStyles";
+import { OptionType } from "../types";
 
 function Input({
   value,
@@ -40,7 +39,7 @@ function MultiInput({
     <Creatable
       isMulti
       menuPortalTarget={document.body}
-      styles={getStyles(darkMode)}
+      styles={selectStyles(darkMode)}
       value={value.map((option) => ({
         value: option,
         label: option,
